ugc/requirements.txt
Flask==2.2.5
flasgger==0.9.7.1
kafka-python==2.0.2
asyncpg==0.30.0
redis==5.0.4
pydantic-settings==2.9.1
aiohttp==3.9.3
pytest==7.4.3
pytest-asyncio==0.21.1 
backoff==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
PyJWT==2.8.0
gunicorn==21.2.0
annotated-types==0.7.0
clickhouse-driver==0.2.9
kafka-python==2.0.2
kafka-python-ng==2.2.3
pytz==2025.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
